refactor: generator 폴더 분리
