+++ src/components/generator/SwitchGenerator.tsx
@@ -3,8 +3,8 @@ import { useRecoilState } from 'recoil';
 import { v4 as uuid } from 'uuid';
 import { componentsState } from 'atoms/components';
 import { TextComponentType, ImgComponentType } from 'types';
-import TextComponentEditor from './TextComponentEditor';
-import ImgComponentEditor from './ImgComponentEditor';
+import TextComponentEditor from './editor/TextComponentEditor';
+import ImgComponentEditor from './editor/ImgComponentEditor';
 
 const SwitchGenerator = () => {
   const [components, setComponents] = useRecoilState(componentsState);
