+++ src/components/generator/editor/EditorComponent.tsx
