+++ src/components/generator/editor/ImgComponentEditor.tsx
