+++ src/components/generator/editor/TextComponentEditor.tsx
